Move 2 remote request classes into subnamespace
